split save

+++ packages/db/src/index.ts
@@ -22,6 +22,8 @@ export * from './schema/typeDef.js'
 export * from './modify/modify.js'
 
 const SCHEMA_FILE = 'schema.json'
+const COMMON_SDB_FILE = 'common.sdb'
+const block_sdb_file = (typeId: number, start: number) => `${typeId}_${start}.sdb`
 
 export class BasedDb {
   isDraining: boolean = false
@@ -118,11 +120,11 @@ export class BasedDb {
     } catch (err) {}
     if (this.splitDump) {
       const dumps = (await fs.readdir(this.fileSystemPath)).filter(
-        (fname: string) => fname.endsWith('.sdb') && fname != 'common.sdb',
+        (fname: string) => fname.endsWith('.sdb') && fname != COMMON_SDB_FILE,
       )
 
       this.dbCtxExternal = db.start(this.fileSystemPath, null, false, this.id)
-      db.loadCommon(join(this.fileSystemPath, 'common.sdb'), this.dbCtxExternal)
+      db.loadCommon(join(this.fileSystemPath, COMMON_SDB_FILE), this.dbCtxExternal)
       dumps.forEach((fname) =>
         db.loadRange(join(this.fileSystemPath, fname), this.dbCtxExternal),
       )
@@ -271,15 +273,38 @@ export class BasedDb {
   }
 
   async save() {
-    const dumppath = join(this.fileSystemPath, 'data.sdb')
-    await fs.rm(dumppath).catch(() => {})
-    var rdy = false
-    const pid = this.native.save(dumppath, this.dbCtxExternal)
-    while (!rdy) {
-      await wait(100)
-      if (this.native.isSaveReady(pid, dumppath)) {
-        rdy = true
-        break
+    if (this.splitDump) {
+      let err: number
+
+      err = this.native.saveCommon(join(this.fileSystemPath, COMMON_SDB_FILE), this.dbCtxExternal)
+      console.error('common err', err)
+
+      for (const key in this.schemaTypesParsed) {
+        const def = this.schemaTypesParsed[key]
+        const [_total, lastId] = this.native.getTypeInfo(
+          def.id,
+          this.dbCtxExternal,
+        )
+        const step = 10000 // TODO Make configurable/variable
+
+        for (let start = 0; start < lastId; start += step) {
+          const end = start + step - 1
+          const path = join(this.fileSystemPath, block_sdb_file(def.id, start))
+          err = this.native.saveRange(path, this.schema.types.user.id, start, end, this.dbCtxExternal)
+          console.error('err', err)
+        }
+      }
+    } else {
+      const dumppath = join(this.fileSystemPath, 'data.sdb')
+      await fs.rm(dumppath).catch(() => {})
+      var rdy = false
+      const pid = this.native.save(dumppath, this.dbCtxExternal)
+      while (!rdy) {
+        await wait(100)
+        if (this.native.isSaveReady(pid, dumppath)) {
+          rdy = true
+          break
+        }
       }
     }
   }
